requirement.txt
icecream
lovely-numpy
tqdm
scikit-learn==1.0.2
scikit-optimize
shap
pandas==1.3.5
numpy==1.21.6
matplotlib
seaborn
pysnooper
pyforest
GPy
GPyOpt
xgboost==1.6.2
scipy==1.4.1
umap-learn
adjustText
openpyxl
fire
biopython
